Added SQL startup scripts for in-memory database (H2) Cleanup code Changed id to Long type

+++ SmartCity Core/src/main/java/be/uantwerpen/sc/controllers/TrafficLightController.java
@@ -27,7 +27,7 @@ public class TrafficLightController
     }
 
     @RequestMapping(value = "{id}",method = RequestMethod.GET)
-    public TrafficLight getBot(@PathVariable("id") int id)
+    public TrafficLight getBot(@PathVariable("id") Long id)
     {
         return trafficLightControlService.getTrafficLight(id);
     }
