+++ SmartCity Core/src/main/java/be/uantwerpen/sc/repositories/LinkRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
  * Created by Thomas on 27/02/2016.
  */
 @Repository
-public interface LinkRepository extends CrudRepository<Link, Integer>
+public interface LinkRepository extends CrudRepository<Link, Long>
 {
     List<Link> findAll();
 }
